requirements.txt
flask
pymongo
dotenv
google-genai